alteração do tipo de valor

+++ src/main/java/Zeus/API/ZEUS/Dto/DadosAtualizacaoRacao.java
@@ -2,6 +2,7 @@ package Zeus.API.ZEUS.Dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Date;
 
@@ -9,7 +10,7 @@ public record DadosAtualizacaoRacao(
         Long id,
         String nome,
         int kgQuantidade,
-        int valorPago,
+        double valorPago,
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
         LocalDate dataCompra
 ) {
