+++ src/main/java/Zeus/API/ZEUS/Dto/DadosCadastrosRacao.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Date;
@@ -14,7 +15,7 @@ public record DadosCadastrosRacao (
         @NotNull
         int kgQuantidade,
         @NotNull
-        int valorPago,
+        double valorPago,
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd/MM/yyyy")
         LocalDate dataCompra
 ){
